Added messaging infrastructure to determine when callback halts
Callback halts when the connection_error indicator attribute on the client logic
resets to its default state of false.

A check for the current state was added to the previous address dial routine.
Similar checks will be added to each phase of the callback process.

The send loop was modified so that it can handle multiple connections at once by
adding a pointer to the connection a message should be sent on.

When the handler for sident_response message types is added, it will include a
check for the error indicators state at the beginning so that if it has already
been resolved by a previous message it will simply halt.